feat(dataset): add data streaming functionalities
Python Serial Streamer: tools/stream_dataset.py
Successfully validated header, sent samples (got 0x2e response = Arduino output collision)

Arduino Dataset Streaming Code
File: core/core.ino
Compiles on RP2350 without errors

Closes #6
