+++ core/platform_rp2350.cpp
@@ -1,29 +1,40 @@
+/**
+ * @file platform_rp2350.cpp
+ * @brief RP2350 (Pico 2 W) platform implementation
+ */
+
+#include "config.h"
+
 #ifdef PLATFORM_RP2350
 
 #include <WiFi.h>
 #include <LittleFS.h>
-#include "config.h"
 
 void platform_init() {
   pinMode(LED_PIN, OUTPUT);
 
   #ifdef HAS_WIFI
+  Serial.print("WiFi connecting");
   WiFi.begin(WIFI_SSID, WIFI_PASSWORD);
-  Serial.print("WiFi");
   while (WiFi.status() != WL_CONNECTED) {
     delay(500);
     Serial.print(".");
   }
-  Serial.printf("\nConnected: %s\n", WiFi.localIP().toString().c_str());
+  Serial.printf("\nWiFi connected: %s\n", WiFi.localIP().toString().c_str());
   #endif
 
-  LittleFS.begin();
-  Serial.println("Storage ready (LittleFS)");
+  if (LittleFS.begin()) {
+    Serial.println("Storage ready (LittleFS)");
+  } else {
+    Serial.println("WARNING: LittleFS mount failed");
+  }
 }
 
 void platform_loop() {
   #ifdef HAS_WIFI
-  if (WiFi.status() != WL_CONNECTED) WiFi.reconnect();
+  if (WiFi.status() != WL_CONNECTED) {
+    WiFi.begin(WIFI_SSID, WIFI_PASSWORD);
+  }
   #endif
 }
 
@@ -36,4 +47,4 @@ void platform_blink(uint8_t times) {
   }
 }
 
-#endif
+#endif  // PLATFORM_RP2350
